+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ smaht-portal
 Change Log
 ----------
 
+0.99.2
+======
+`PR 242: BM colo829 truth set <https://github.com/smaht-dac/smaht-portal/pull/242>`_
+
+* Add COLO829 Truth Set Tab to benchmarking page
+* Add a way to change facets and columns from BenchmarkingDataMap
+
+
 0.99.1
 ======
 `PR264: SN remove tissue from cell culture  <https://github.com/smaht-dac/smaht-portal/pull/264>`_